Abstract some remote relative server spec into McpServerRemoteServiceConfig.
